Still working towards async transfer
the device is being lost further into the program, meaning i fixed a few issues. I have more to fix.

I believe the current issue is synchronization, which TransferCommandManager was made to address.
TransitionManager is being adjusted to take a copy of the barrier instead of reconstructing it with given data, which seems obvious in hindsight.

I'm also shifting towards a more data oriented approach rather than object oriented. I'll probably write a big paper about this when I'm done.

Currently, GenerateMips needs to be called from the graphics thread. Currently I'm envisioning a GraphicsManager class, similar to TransitionManager, but a bit simpler since all I currently need from it is calling the GenerateMips function on images. It might be an extension to TransitionManager instead of it's own full class.